ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.25.2
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0

# AI関連ライブラリ
torch==2.2.0
torchvision==0.17.0
sentence-transformers>=3.0.0
onnxruntime==1.16.3

# AWS関連
boto3==1.29.0
botocore==1.32.0

# ベクトル検索
faiss-cpu==1.7.4